work in progress

+++ HWInterface/MPA2Interface.h
@@ -116,7 +116,7 @@ class MPA2Interface : public ReadoutChipInterface
     bool                 configPixel(Ph2_HwDescription::Chip* pChip, std::string cReg, uint16_t row, uint16_t col, uint8_t pValue, bool pVerify);
     uint16_t             readPixel(Ph2_HwDescription::Chip* pChip, std::string cReg, uint16_t row, uint16_t col);
     bool                 setVrefFromFuseID(Ph2_HwDescription::ReadoutChip* pMPA2)  override;
-    bool                 setVref(Ph2_HwDescription::ReadoutChip* pMPA2, uint8_t theVrefRegisterValue)  override;
+    bool                 setVref(Ph2_HwDescription::ReadoutChip* pMPA2, uint16_t theVrefRegisterValue)  override;
     float                ADCMeasure(Ph2_HwDescription::Chip* pMPA2, uint32_t nreads = 5);
     bool                 selectBlock(Ph2_HwDescription::Chip* pMPA2, uint8_t block, uint8_t testPoint = 0, uint8_t swEn = 0);
     uint32_t             readADCGround(Ph2_HwDescription::ReadoutChip* pMPA2);
